+++ NEWS.md
@@ -108,8 +108,8 @@ Red Panda C++ Version 2.27
   - enhancement: By default, use monospaced font to display register values in the CPU Info dialog.
   - fix: Negative values in register like AH/AL are wrongs displayed as 32/64-bit number.
   - Change: Change background color for highlighted buttons in the default theme.
-  - enhancement: Make matched contents more obvious in the code suggestion popup.
-  - enhancement: Make matched contents more obvious in the header suggestion popup.
+  - enhancement: Make colors in code suggestion popup consistent with the editor.
+  - enhancement: Make colors in header suggestion popup consistent with the editor.
     
 Red Panda C++ Version 2.26
   - enhancement: Code suggestion for embedded std::vectors.